requirements.txt
annotated-types
anyio
asgiref
certifi
charset-normalizer
contourpy
cycler
datetime-truncate
Django
django-admin-charts
django-admin-interface
django-bower
django-colorfield
django-jazzmin
django-memoize
django-multiselectfield
django-nvd3
fonttools
h11
httpcore
httpx
idna
Jinja2
kiwisolver
langsmith
MarkupSafe
matplotlib
numpy
orjson
packaging
pandas
pillow
pydantic
pydantic_core
pyparsing
python-dateutil
python-nvd3
python-slugify
pytz
requests
requests-toolbelt
seaborn
six
sniffio
sqlparse
text-unidecode
typing-inspection
typing_extensions
tzdata
urllib3
zstandard